Add aliases for integer types
Added a typedef 'uintsize' for an unsigned integer type with a width of
at least 16 bits that is large enough to store the size of any array or
object.

Added the following typedefs for standard integer types:
 - sbyte    signed char
 - byte     unsigned char
 - ushort   unsigned short
 - uint     unsigned int
 - ulong    unsigned long

Added the following typedefs for integers with a certain number of bits
 - intN         signed integer with exactly N bits
 - intN_least   smallest signed integer with at least N bits
 - intN_fast    fastest signed integer with at least N bits
 - uintN        unsigned integer with exactly N bits
 - uintN_least  smallest unsigned integer with at least N bits
 - uintN_fast   fastest unsigned integer with at least N bits
for the follwing N: 8, 16, 32, 64.

The types with an exact with and with a width of 64 bits or greater are
optional and may not be defined on certain platforms.