Disable Visual Studio warning C4479
Disabled Visual Studio warning C4479:
    nonstandard extension used: literal suffix

Because the Visual Studio is configured to compile with the /TC
(Compile As C) and /Za (Disable Language Extensions) options, the
compiler will issue a warning whenever code does not conform to the C89
standard.

However, some headers use a predefined macro to determine if the
compiler supports and provides the C99 stdint.h header before using it.
The use of certain macros defined by stdint.h cause the C4479 warning.

Since these macros are only used when it has been determined that they
are supported by the compiler, this warning can safely be ignored.