Update doxygen config
Excluded symbol CFLAT_BOOL from the output because it was only used
within file scope and then undefined with #undef.

Set the value of MAX_INITIALIZER_LINES to 0 in order to hide the values
of macros, since some values are platform dependant and could lead to
incorrect assumptions based on the generated documentation.