Add reference counting functions
Added the following functions to allocate/deallocate reference counted
memory blocks:
 - Object_New()
 - Object_Delete()

Added the following functions for incrementing/decrementing the
reference count of a memory block:
 - Object_Aquire()
 - Object_Release()

The reference counted memory can be used to hold any object so that it
becomes reference counted. A 'destructor' callback can be set with
Object_New() so that the object can be cleaned up before the memory
is deallocated.