Add StringHelper and move String struct definition
The definition of the String struct was moved to a header
String-private.h in the src/ folder so that the definition is available
within the CFlat.Core but not outside.

In doing so, it is possible to allocate instances of String on the
stack, which allows efficient wrapping of regular C-strings.