Wrap variable arguments handling
In some functions it is neccesary to pass a va_list by reference.
However, when va_list is declared as an array type this will result in a
compiler warning. To resolve this, the va_list must be encompassed by a
struct.

The va_start, va_end and va_arg macros were also wrapped for consistency.